copy updated .spec

+++ micq.spec
@@ -1,18 +1,17 @@
-%define name		micq
-%define version	0.4.10.4
-%define release	1
-
-Summary:		text/line based ICQ client with many features
-Name:			%{name}
-Version:		%{version}
-Release:		%{release}
-Source:			%{name}-%{version}.tgz
+Summary:		text/line based ICQ client with many features%{!?_with_tcl: [no Tcl]}%{!?_with_ssl: [no SSL]}
+Name:			micq%{!?_with_tcl:-notcl}%{!?_with_ssl:-nossl}
+Version:		0.4.11
+Release:		1
+Source:			micq-%{version}.tgz
 URL:			http://www.micq.org/
 Group:			Networking/ICQ
 Packager:		Rüdiger Kuhlmann <info@ruediger-kuhlmann.de>
 License:		GPL-2
-BuildRoot:		%{_tmppath}/build-%{name}-%{version}
+BuildRoot:		%{_tmppath}/build-micq-%{version}
 Prefix:			%{_prefix}
+
+%{?_with_ssl:BuildRequires: openssl-devel}
+%{?_with_tcl:BuildRequires: tcl-devel}
 
 %description
 mICQ is a portable, small, yet powerful console based ICQ client. It
@@ -32,6 +31,12 @@ Authors: Matthew D. Smith (deceased)
          Rüdiger Kuhlmann <info@ruediger-kuhlmann.de>
 
 %changelog
+* Sat Jan 17 2004 Rüdiger Kuhlmann <info@ruediger-kuhlmann.de>
+- new upstream release 0.4.11
+
+* Mon Oct  6 2003 Rüdiger Kuhlmann <info@ruediger-kuhlmann.de>
+- new upstream bug fix release 0.4.10.5
+
 * Mon Sep 22 2003 Rüdiger Kuhlmann <info@ruediger-kuhlmann.de>
 - new upstream bug fix release 0.4.10.4
 - fixes remote DoS
@@ -61,39 +66,49 @@ Authors: Matthew D. Smith (deceased)
 - first RPM
 
 %prep
-rm -rf $RPM_BUILD_ROOT
+test $RPM_BUILD_ROOT != / && rm -rf $RPM_BUILD_ROOT
 
-%setup -q -n %{name}-%{version}
+%setup -q -n micq-%{version}
 
 %build
-%configure --disable-dependency-tracking --enable-tcl --enable-ssl CFLAGS=-O4
+%configure --disable-dependency-tracking CFLAGS=-O4 \
+	%{?_with_tcl:--enable-tcl}%{!?_with_tcl:--disable-tcl} \
+	%{?_with_ssl:--enable-ssl}%{!?_with_ssl:--disable-ssl}
 make
 
 %install
 
 make install DESTDIR=$RPM_BUILD_ROOT
-%{__mkdir_p} $RPM_BUILD_ROOT/%{_libdir}/menu
-cat << EOF > $RPM_BUILD_ROOT/%{_libdir}/menu/micq
+%if %{?update_menus:1}%{!?update_menus:0}
+%{__mkdir_p} $RPM_BUILD_ROOT%{_libdir}/menu
+cat << EOF > $RPM_BUILD_ROOT%{_libdir}/menu/micq
 ?package(micq):needs=text section=Networking/ICQ \
   title="mICQ" command="%{_bindir}/micq" hints="ICQ client"\
   icon=%{_datadir}/pixmaps/micq.xpm
 EOF
-install -D -m 644 doc/micq.xpm $RPM_BUILD_ROOT/%{_datadir}/pixmaps/micq.xpm
+install -D -m 644 -p doc/micq.xpm $RPM_BUILD_ROOT%{_datadir}/pixmaps/micq.xpm
+%endif
 
 %clean
-rm -rf "${RPM_BUILD_ROOT}"
+test $RPM_BUILD_ROOT != / && rm -rf $RPM_BUILD_ROOT
 
 %files
 %defattr(-,root,root,0755)
 %doc NEWS AUTHORS FAQ README TODO COPYING COPYING-GPLv2
-%doc doc/README.i18n doc/README.logformat doc/icq091.txt doc/icqv7.txt
+%doc doc/README.i18n doc/README.logformat doc/README.ssl
 %{_bindir}/*
 %{_datadir}/micq
+%if %{?update_menus:1}%{!?update_menus:0}
 %{_libdir}/menu/micq
-%{_mandir}/*
+%{_datadir}/pixmaps/micq.xpm
+%endif
+%{_mandir}/man?/*
+%{_mandir}/*/man?/*
 
+%if %{?update_menus:1}%{!?update_menus:0}
 %post
 %{update_menus} || true
 
 %postun
 %{clean_menus} || true
+%endif
